fixed fixtures
I chose to use Cornell MARCXML records for a lot of the fixtures.
Unfortunately, the Cornell MARCXML I used contained the incorrect 003 field
value (the OCLC code 'COO' instead of MARC code 'NIC').  When I added the
assertion to the Marcxml class to test that the 003 value was in the allowed
set a lot of tests started failing. Ironically, this was not because the code
was bad, but because the code was working! The code was catching the fact that
the test MARCXML files contained an invalid 003 field value.  This commit
updates the fixtures with the correct Cornell 003 value: NIC.