Change script to be compatible with git v1.7.1
This commit is a bug fix. The previous version of the script relied on
the "%cI" (strict ISO 8601 format) git log format placeholder.
Unfortunately, this format placeholder is not available in v1.7.1, which
resulted in an echo of "%cI" instead of the commit date.

The format placeholder was changed to "%ci", which is a non-strict
ISO 8601 format available on v1.7.1.